pyproject.toml
[tool.poetry]
name = "task-10-sql"
version = "0.1.0"
description = ""
authors = ["Sasha Barvynska <alexbarvynska@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.5"
requests = "^2.28.2"
flask-restful = "^0.3.9"
faker = "^18.4.0"
factory-boy = "3.2.1"
psycopg2 = "2.9.6"
jsonpickle = "^3.0.1"
typer = "^0.9.0"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
flask = "2.3.2"
flasgger = "0.9.7.1"
flask-sqlalchemy = "^3.0.3"
sqlalchemy = "^2.0.15"
apispec = "^6.3.0"
flask-apispec = "^0.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
skip=".gitignore,venv"
multi_line_output = 5
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
atomic = true

[tool.flask.logging]
file = "my_app.log"
level = "INFO"
